requirements.txt
discord.py
pynacl
ffmpeg==1.4
yt-dlp
flask
